Extract e2e database metadata helpers
